feat(Entry): add for entry controller authorization

+++ src/JournalBackend.Core/Features/Entries/EntryController.cs
@@ -1,6 +1,8 @@
 using Microsoft.AspNetCore.Mvc;
 using JournalBackend.Models;
 using JournalBackend.Services;
+using Microsoft.Extensions.Configuration.UserSecrets;
+using Microsoft.AspNetCore.Authorization;
 
 namespace JournalBackend.Controllers;
 
@@ -16,41 +18,75 @@ public class EntryController : ControllerBase
         _logger = logger;
     }
 
+    [Authorize]
     [HttpGet]
     public async Task<ActionResult<List<Entry>>> GetAll(JournalDbContext db)
     {
-        var allEntries = await EntryService.GetAll(db);
-        return Ok(allEntries);
+        string? role = User.Claims.FirstOrDefault(claim => claim.Type.Equals("role"))?.Value;
+        
+        if (role == "admin") {
+          var allEntries = await EntryService.GetAll(db);
+          return Ok(allEntries);
+        } else {
+            return Unauthorized();
+        }
     }
 
+    [Authorize]
     [HttpGet("User/{userId}")]
     public async Task<ActionResult<List<Entry>>> GetAllOfUser(JournalDbContext db, string userId)
     {
+        string? authUserId = User.Claims.FirstOrDefault(claim => claim.Type.Equals("sub", StringComparison.OrdinalIgnoreCase))?.Value;
+
+        if (userId != authUserId) {
+            return Unauthorized();
+        }
+
         var allUserEntries = await EntryService.GetAllOfUser(db, userId);
         return Ok(allUserEntries);
     }
 
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<ActionResult<Entry>> Get(JournalDbContext db, string id)
     {
+        string? userId = User.Claims.FirstOrDefault(claim => claim.Type.Equals("sub", StringComparison.OrdinalIgnoreCase))?.Value;
+        
         var entry = await EntryService.Get(db, id);
 
-        if (entry is null)
+        if (entry is null){
             return NotFound();
-        
+        }
+
+        if (entry.UserId != userId) {
+            return Unauthorized();
+        }
+
         return Ok(entry);
     }
 
+    [Authorize]
     [HttpPost]
     public IActionResult Create(JournalDbContext db, Entry entry)
     {
+        // Access userId 
+        string? userId = User.Claims.FirstOrDefault(claim => claim.Type.Equals("sub", StringComparison.OrdinalIgnoreCase))?.Value;
+
+        if (userId is null) {
+            return Unauthorized();
+        }
+
+        entry.UserId = userId;
         EntryService.Add(db, entry);
         return CreatedAtAction(nameof(Get), new { entry.Id }, entry);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(JournalDbContext db, string id, Entry entry)
     {
+        string? userId = User.Claims.FirstOrDefault(claim => claim.Type.Equals("sub", StringComparison.OrdinalIgnoreCase))?.Value;
+
         if (id != entry.Id)
             return BadRequest();
         
@@ -58,18 +94,28 @@ public class EntryController : ControllerBase
         if (existingEntry is null)
             return NotFound();
 
+        if (entry.UserId != userId) {
+            return Unauthorized();
+        }
+
         EntryService.Update(db, entry);
 
         return NoContent();
     }
 
-
+    [Authorize]
     [HttpDelete("{id}")]
-    public IActionResult Delete(JournalDbContext db, string id)
+    public async Task<IActionResult> Delete(JournalDbContext db, string id)
     {
-        var entry = EntryService.Get(db, id);
+        string? userId = User.Claims.FirstOrDefault(claim => claim.Type.Equals("sub", StringComparison.OrdinalIgnoreCase))?.Value;
+
+        var entry = await EntryService.Get(db, id);
         if (entry is null) 
             return NotFound();
+
+        if (entry.UserId != userId) {
+            return Unauthorized();
+        }
 
         EntryService.Delete(db, id);
         return NoContent();
